busiaclone/busiav2/requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
requests==2.32.3
Pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
